Make ansible-lint and yamllint use more strict rules.
This patch modifies configuration of both ansible-lint and yamllint
to check for more rules, resulting in a more strict verification.

For ansible-lint verification of errors E301 and E502 are skipped.

\Ansible-lint issues E301 for shel tasks, complaining that things
that do not need to be changed should do nothing, even if the script
does not change anything and just gather informattion.

All tasks should have a name (E502), but this would change almost
all example playbooks, and this change should be adreesed on its
own patch, if this rule is to be used for the example playbooks.

Yamllint disabled rules (comments, comments-indentation and
indentation) would introduce a huge amount of small changes, and
are left for future changes, if deemed necessary.